Extract EnhancedGraphTable(MouseAdapter|PopupMenuListener) classes to their own files
