+++ frontend/ui/impl/src/main/java/com/virtuslab/gitmachete/frontend/ui/impl/table/EnhancedGraphTable.java
@@ -2,36 +2,25 @@ package com.virtuslab.gitmachete.frontend.ui.impl.table;
 
 import static com.intellij.openapi.application.ModalityState.NON_MODAL;
 import static com.virtuslab.gitmachete.frontend.datakeys.DataKeys.typeSafeCase;
-import static com.virtuslab.gitmachete.frontend.defs.ActionIds.CHECK_OUT;
 import static com.virtuslab.gitmachete.frontend.defs.ActionIds.OPEN_MACHETE_FILE;
 import static com.virtuslab.gitmachete.frontend.resourcebundles.GitMacheteBundle.getString;
 import static io.vavr.API.$;
 import static io.vavr.API.Case;
 import static io.vavr.API.Match;
 
-import java.awt.Point;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Timer;
-import java.util.TimerTask;
 import java.util.function.Consumer;
 
-import javax.swing.JPopupMenu;
 import javax.swing.ListSelectionModel;
-import javax.swing.SwingUtilities;
-import javax.swing.event.PopupMenuEvent;
 
 import com.intellij.ide.DataManager;
 import com.intellij.notification.Notification;
 import com.intellij.notification.NotificationAction;
 import com.intellij.notification.NotificationType;
-import com.intellij.openapi.actionSystem.ActionGroup;
 import com.intellij.openapi.actionSystem.ActionManager;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.CommonDataKeys;
-import com.intellij.openapi.actionSystem.DataContext;
 import com.intellij.openapi.actionSystem.DataProvider;
 import com.intellij.openapi.actionSystem.Presentation;
 import com.intellij.openapi.application.ApplicationManager;
@@ -43,7 +32,6 @@ import com.intellij.openapi.vfs.VirtualFileManager;
 import com.intellij.openapi.vfs.newvfs.BulkFileListener;
 import com.intellij.openapi.vfs.newvfs.events.VFileContentChangeEvent;
 import com.intellij.openapi.vfs.newvfs.events.VFileEvent;
-import com.intellij.ui.PopupMenuListenerAdapter;
 import com.intellij.ui.ScrollingUtil;
 import com.intellij.util.ModalityUiUtil;
 import com.intellij.util.messages.Topic;
@@ -51,6 +39,7 @@ import com.intellij.util.ui.JBUI;
 import git4idea.repo.GitRepository;
 import git4idea.repo.GitRepositoryChangeListener;
 import io.vavr.collection.Set;
+import lombok.AccessLevel;
 import lombok.CustomLog;
 import lombok.Getter;
 import lombok.Setter;
@@ -68,10 +57,8 @@ import com.virtuslab.branchlayout.api.readwrite.IBranchLayoutWriter;
 import com.virtuslab.gitmachete.backend.api.IGitMacheteRepositorySnapshot;
 import com.virtuslab.gitmachete.backend.api.NullGitMacheteRepositorySnapshot;
 import com.virtuslab.gitmachete.frontend.datakeys.DataKeys;
-import com.virtuslab.gitmachete.frontend.defs.ActionGroupIds;
 import com.virtuslab.gitmachete.frontend.defs.ActionPlaces;
 import com.virtuslab.gitmachete.frontend.defs.FileTypeIds;
-import com.virtuslab.gitmachete.frontend.graph.api.items.IGraphItem;
 import com.virtuslab.gitmachete.frontend.graph.api.repository.IRepositoryGraph;
 import com.virtuslab.gitmachete.frontend.graph.api.repository.IRepositoryGraphCache;
 import com.virtuslab.gitmachete.frontend.graph.api.repository.NullRepositoryGraph;
@@ -90,7 +77,6 @@ import com.virtuslab.qual.guieffect.UIThreadUnsafe;
  *  data like last clicked branch name, opened project or {@link IGitMacheteRepositorySnapshot} of current
  *  repository for actions.
  */
-// TODO (#99): consider applying SpeedSearch for branches and commits
 @ExtensionMethod({GitMacheteBundle.class, GitVfsUtils.class})
 @CustomLog
 public final class EnhancedGraphTable extends BaseEnhancedGraphTable
@@ -113,6 +99,8 @@ public final class EnhancedGraphTable extends BaseEnhancedGraphTable
   @UIEffect
   private @Nullable IGitMacheteRepositorySnapshot gitMacheteRepositorySnapshot;
 
+  @Getter(AccessLevel.PACKAGE)
+  @Setter(AccessLevel.PACKAGE)
   @UIEffect
   private @Nullable String selectedBranchName;
 
@@ -415,102 +403,4 @@ public final class EnhancedGraphTable extends BaseEnhancedGraphTable
         Case($(), (Object) null));
   }
 
-  private static class EnhancedGraphTableMouseAdapter extends MouseAdapter {
-    private final EnhancedGraphTable graphTable;
-    private final EnhancedGraphTablePopupMenuListener popupMenuListener;
-
-    @UIEffect
-    EnhancedGraphTableMouseAdapter(EnhancedGraphTable graphTable) {
-      this.graphTable = graphTable;
-      this.popupMenuListener = new EnhancedGraphTablePopupMenuListener(graphTable);
-    }
-
-    @Override
-    @UIEffect
-    public void mouseClicked(MouseEvent e) {
-      Point point = e.getPoint();
-      int row = graphTable.rowAtPoint(point);
-      int col = graphTable.columnAtPoint(point);
-
-      // check if we click on one of the branches
-      if (row < 0 || col < 0) {
-        return;
-      }
-
-      BranchOrCommitCell cell = (BranchOrCommitCell) graphTable.getModel().getValueAt(row, col);
-      IGraphItem graphItem = cell.getGraphItem();
-      if (!graphItem.isBranchItem()) {
-        return;
-      }
-
-      graphTable.selectedBranchName = graphItem.asBranchItem().getBranch().getName();
-      performActionAfterChecks(e, point);
-    }
-
-    @UIEffect
-    private void performActionAfterChecks(MouseEvent e, Point point) {
-      ActionManager actionManager = ActionManager.getInstance();
-      if (SwingUtilities.isRightMouseButton(e) || isCtrlClick(e)) {
-        ActionGroup contextMenuActionGroup = (ActionGroup) actionManager.getAction(ActionGroupIds.CONTEXT_MENU);
-        val actionPopupMenu = actionManager.createActionPopupMenu(ActionPlaces.CONTEXT_MENU, contextMenuActionGroup);
-        JPopupMenu popupMenu = actionPopupMenu.getComponent();
-        popupMenu.addPopupMenuListener(popupMenuListener);
-        popupMenu.show(graphTable, (int) point.getX(), (int) point.getY());
-      } else if (SwingUtilities.isLeftMouseButton(e) && e.getClickCount() == 2 && !e.isConsumed()) {
-
-        val gitMacheteRepositorySnapshot = graphTable.gitMacheteRepositorySnapshot;
-        if (gitMacheteRepositorySnapshot != null) {
-          val currentBranchIfManaged = gitMacheteRepositorySnapshot
-              .getCurrentBranchIfManaged();
-          val isSelectedEqualToCurrent = currentBranchIfManaged != null
-              && currentBranchIfManaged.getName().equals(graphTable.selectedBranchName);
-          if (isSelectedEqualToCurrent) {
-            return;
-          }
-        }
-
-        e.consume();
-        DataContext dataContext = DataManager.getInstance().getDataContext(graphTable);
-        val actionEvent = AnActionEvent.createFromDataContext(ActionPlaces.CONTEXT_MENU, new Presentation(), dataContext);
-        actionManager.getAction(CHECK_OUT).actionPerformed(actionEvent);
-      }
-    }
-
-    // this method is needed as some macOS users use Ctrl + left-click as a replacement for the right-click
-    @UIEffect
-    private boolean isCtrlClick(MouseEvent e) {
-      return e.isControlDown() && SwingUtilities.isLeftMouseButton(e);
-    }
-  }
-
-  private static class EnhancedGraphTablePopupMenuListener extends PopupMenuListenerAdapter {
-    private final EnhancedGraphTable graphTable;
-
-    @UIEffect
-    EnhancedGraphTablePopupMenuListener(EnhancedGraphTable graphTable) {
-      this.graphTable = graphTable;
-    }
-
-    @Override
-    @UIEffect
-    public void popupMenuWillBecomeVisible(PopupMenuEvent popupMenuEvent) {
-      // This delay is needed to avoid `focus transfer` effect when at the beginning row selection is light-blue,
-      // but when the context menu is created (in a fraction of a second),
-      // selection loses focus on the context menu and becomes dark blue.
-      // TimerTask can't be replaced by lambda because it's not a SAM (single abstract method).
-      // For more details see https://stackoverflow.com/a/37970821/10116324.
-      new Timer().schedule(new TimerTask() {
-        @Override
-        public void run() {
-          ModalityUiUtil.invokeLaterIfNeeded(NON_MODAL, () -> graphTable.setRowSelectionAllowed(true));
-        }
-      }, /* delay in ms */ 35);
-    }
-
-    @Override
-    @UIEffect
-    public void popupMenuWillBecomeInvisible(PopupMenuEvent popupMenuEvent) {
-      graphTable.setRowSelectionAllowed(false);
-    }
-  }
 }
